Refactor preposition subset learner. Introduce MatchedObjectNode
