+++ tests/learner/learner_test.py
@@ -20,6 +20,7 @@ from adam.curriculum.preposition_curriculum import (
 )
 from adam.language_specific.english.english_language_generator import IGNORE_COLORS
 from adam.learner import LearningExample
+from adam.learner.preposition_subset import PrepositionSubsetLanguageLearner
 from adam.learner.subset import SubsetLanguageLearner
 from adam.ontology.phase1_ontology import BALL, LEARNER
 from adam.ontology import OntologyNode
@@ -96,7 +97,7 @@ def test_subset_learner_dog():
 
 
 def test_subset_preposition_on_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     on_train_curriculum = _make_on_training(num_samples=1, noise_objects=False)
     on_test_curriculum = _make_on_tests(num_samples=1, noise_objects=False)
 
@@ -105,9 +106,8 @@ def test_subset_preposition_on_learner():
         linguistic_description,
         perceptual_representation,
     ) in on_train_curriculum.instances():
-        learner.observe_static_prepositions(  # pylint:disable=protected-access
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -115,15 +115,13 @@ def test_subset_preposition_on_learner():
         test_lingustics_description,
         test_perceptual_representation,
     ) in on_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(  # pylint:disable=protected-access
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_lingustics_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
 
 def test_subset_preposition_beside_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     beside_train_curriculum = _make_beside_training(num_samples=1, noise_objects=False)
     beside_test_curriculum = _make_beside_tests(num_samples=1, noise_objects=False)
 
@@ -132,9 +130,8 @@ def test_subset_preposition_beside_learner():
         linguistic_description,
         perceptual_representation,
     ) in beside_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -142,15 +139,13 @@ def test_subset_preposition_beside_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in beside_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
 
 def test_subset_preposition_under_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     under_train_curriculum = _make_under_training(num_samples=1, noise_objects=False)
     under_test_curriculum = _make_under_tests(num_samples=1, noise_objects=False)
 
@@ -159,9 +154,8 @@ def test_subset_preposition_under_learner():
         linguistic_description,
         perceptual_representation,
     ) in under_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -169,15 +163,13 @@ def test_subset_preposition_under_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in under_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
 
 def test_subset_preposition_over_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     over_train_curriculum = _make_over_training(num_samples=1, noise_objects=False)
     over_test_curriculum = _make_over_tests(num_samples=1, noise_objects=False)
 
@@ -186,9 +178,8 @@ def test_subset_preposition_over_learner():
         linguistic_description,
         perceptual_representation,
     ) in over_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -196,9 +187,7 @@ def test_subset_preposition_over_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in over_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
@@ -206,7 +195,7 @@ def test_subset_preposition_over_learner():
 # See https://github.com/isi-vista/adam/issues/422
 @pytest.mark.skip(msg="In Preposition Test Temporarily Disabled")
 def test_subset_preposition_in_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     in_train_curriculum = _make_in_training(num_samples=1, noise_objects=False)
     in_test_curriculum = _make_in_tests(num_samples=1, noise_objects=False)
 
@@ -215,9 +204,8 @@ def test_subset_preposition_in_learner():
         linguistic_description,
         perceptual_representation,
     ) in in_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -225,15 +213,13 @@ def test_subset_preposition_in_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in in_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
 
 def test_subset_preposition_behind_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     behind_train_curriculum = _make_behind_training(num_samples=1, noise_objects=False)
     behind_test_curriculum = _make_behind_tests(num_samples=1, noise_objects=False)
 
@@ -242,9 +228,8 @@ def test_subset_preposition_behind_learner():
         linguistic_description,
         perceptual_representation,
     ) in behind_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -252,15 +237,13 @@ def test_subset_preposition_behind_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in behind_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
 
 
 def test_subset_preposition_in_front_learner():
-    learner = SubsetLanguageLearner()
+    learner = PrepositionSubsetLanguageLearner()
     in_front_train_curriculum = _make_in_front_training(
         num_samples=1, noise_objects=False
     )
@@ -271,9 +254,8 @@ def test_subset_preposition_in_front_learner():
         linguistic_description,
         perceptual_representation,
     ) in in_front_train_curriculum.instances():
-        learner.observe_static_prepositions(
-            PerceptionGraph.from_frame(perceptual_representation.frames[0]),
-            linguistic_description,
+        learner.observe(
+            LearningExample(perceptual_representation, linguistic_description)
         )
 
     for (
@@ -281,8 +263,6 @@ def test_subset_preposition_in_front_learner():
         test_linguistic_description,
         test_perceptual_representation,
     ) in in_front_test_curriculum.instances():
-        descriptions_from_learner = learner.describe_preposition(
-            PerceptionGraph.from_frame(test_perceptual_representation.frames[0])
-        )
+        descriptions_from_learner = learner.describe(test_perceptual_representation)
         gold = test_linguistic_description.as_token_sequence()
         assert [desc.as_token_sequence() for desc in descriptions_from_learner][0] == gold
